vm_creator/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
psycopg2-binary